amplify/functions/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
boto3==1.35.76
mangum==0.17.0
python-jose[cryptography]==3.3.0
pwdlib[argon2]==0.2.1
python-multipart==0.0.20
